feat:fine

+++ src/main/java/org/pizza/java/spring_la_mia_pizzeria_webapi/controller/PizzaRestController.java
@@ -1,10 +1,13 @@
 package org.pizza.java.spring_la_mia_pizzeria_webapi.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.pizza.java.spring_la_mia_pizzeria_webapi.model.Pizza;
 import org.pizza.java.spring_la_mia_pizzeria_webapi.repository.PizzaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,45 +18,54 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import org.springframework.web.bind.annotation.RestController;
 
-
-
 @RestController
-
 @RequestMapping("/api/pizzas")
 public class PizzaRestController {
+
     @Autowired
-  private PizzaRepository pizzaRepository;
-  
-  
+    private PizzaRepository pizzaRepository;
+
     @GetMapping
-public List<Pizza> index() {
-    return pizzaRepository.findAll();
-}
+    public List<Pizza> index() {
+        List<Pizza> pizzas = pizzaRepository.findAll();
+        return pizzas;
+    }
 
-@GetMapping("{id}")
-public Pizza get(@PathVariable Integer id) { 
-    return pizzaRepository.findById(id).get();
-}
+    @GetMapping("/{id}")
+    public ResponseEntity<Pizza> show(@PathVariable Integer id) {
+        Optional<Pizza> pizzaAttempt = pizzaRepository.findById(id);
 
-@PostMapping
-public Pizza store(@RequestBody Pizza pizza){ 
-   
-    return pizzaRepository.save(pizza);
-}
+        if (pizzaAttempt.isEmpty()) {
+            return new ResponseEntity<Pizza>(HttpStatus.NOT_FOUND);
+        }
 
-@PutMapping ("{id}")
-public Pizza update( @RequestBody Pizza pizza , @PathVariable Integer id) {
-    pizza.setId(id); 
-    return pizzaRepository.save(pizza);
+        return new ResponseEntity<Pizza>(pizzaAttempt.get(), HttpStatus.OK);
+    }
 
- }
+    @PostMapping
+    public ResponseEntity<Pizza> store(@RequestBody Pizza pizza) {
+        return new ResponseEntity<Pizza>(pizzaRepository.save(pizza), HttpStatus.OK);
+    }
 
-@DeleteMapping("/{id}")
-public void delete(@PathVariable Integer id) { 
-    pizzaRepository.deleteById(id);
-}
+    @PostMapping("/{id}")
+    public ResponseEntity<Pizza> update(@RequestBody Pizza pizza, @PathVariable Integer id) {
 
+        if (pizzaRepository.findById(id).isEmpty()) {
+            return new ResponseEntity<Pizza>(HttpStatus.NOT_FOUND);
 
+        }
 
+        pizza.setId(id);
+        return new ResponseEntity<Pizza>(pizzaRepository.save(pizza), HttpStatus.OK);
+    }
 
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Pizza> delete(@PathVariable Integer id) {
+        if (pizzaRepository.findById(id).isEmpty()) {
+            return new ResponseEntity<Pizza>(HttpStatus.NOT_FOUND);
+        }
+
+        pizzaRepository.deleteById(id);
+        return new ResponseEntity<Pizza>(HttpStatus.OK);
+    }
 }
